made the file application when created class Wrapper for all method with __init__ method
